requirements.txt
numpy~=2.1.3
matplotlib~=3.10.0
scipy~=1.14.1
colour-science~=0.4.6
scikit-image~=0.25.0
dotmap~=1.3.30
opt-einsum~=3.4.0
napari~=0.5.5
magicgui~=0.10.0
lmfit~=1.3.2
pyqt5~=5.15.9
numba~=0.61.0
OpenImageIO~=3.0.3.1
pyopencl
joblib
siphash24